+++ src/payments/lazy.js
@@ -1,30 +1,32 @@
-function prop (object, name, f) {
-  Object.defineProperty(object, name, {
-    configurable: true,
-    enumerable: true,
-    get: function () {
-      let value = f.call(this)
-      this[name] = value
-      return value
-    },
-    set: function (value) {
-      Object.defineProperty(this, name, {
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+function prop(object, name, f) {
+    Object.defineProperty(object, name, {
         configurable: true,
         enumerable: true,
-        value: value,
-        writable: true
-      })
-    }
-  })
+        get() {
+            const _value = f.call(this);
+            this[name] = _value;
+            return _value;
+        },
+        set(_value) {
+            Object.defineProperty(this, name, {
+                configurable: true,
+                enumerable: true,
+                value: _value,
+                writable: true,
+            });
+        },
+    });
 }
-
-function value (f) {
-  let value
-  return function () {
-    if (value !== undefined) return value
-    value = f()
-    return value
-  }
+exports.prop = prop;
+function value(f) {
+    let _value;
+    return () => {
+        if (_value !== undefined)
+            return _value;
+        _value = f();
+        return _value;
+    };
 }
-
-module.exports = { prop, value }
+exports.value = value;