+++ src/templates/pubkey/output.js
@@ -1,15 +1,15 @@
+"use strict";
 // {pubKey} OP_CHECKSIG
-
-const bscript = require('../../script')
-const OPS = require('bitcoin-ops')
-
-function check (script) {
-  const chunks = bscript.decompile(script)
-
-  return chunks.length === 2 &&
-    bscript.isCanonicalPubKey(chunks[0]) &&
-    chunks[1] === OPS.OP_CHECKSIG
+Object.defineProperty(exports, "__esModule", { value: true });
+const bscript = require("../../script");
+const script_1 = require("../../script");
+function check(script) {
+    const chunks = bscript.decompile(script);
+    return (chunks.length === 2 &&
+        bscript.isCanonicalPubKey(chunks[0]) &&
+        chunks[1] === script_1.OPS.OP_CHECKSIG);
 }
-check.toJSON = function () { return 'pubKey output' }
-
-module.exports = { check }
+exports.check = check;
+check.toJSON = () => {
+    return 'pubKey output';
+};