+++ test/block.js
@@ -1,6 +1,6 @@
 const { describe, it, beforeEach } = require('mocha')
 const assert = require('assert')
-const Block = require('../src/block')
+const Block = require('..').Block
 
 const fixtures = require('./fixtures/block')
 
@@ -32,6 +32,9 @@ describe('Block', function () {
         assert.strictEqual(block.version, f.version)
         assert.strictEqual(block.prevHash.toString('hex'), f.prevHash)
         assert.strictEqual(block.merkleRoot.toString('hex'), f.merkleRoot)
+        if (block.witnessCommit) {
+          assert.strictEqual(block.witnessCommit.toString('hex'), f.witnessCommit)
+        }
         assert.strictEqual(block.timestamp, f.timestamp)
         assert.strictEqual(block.bits, f.bits)
         assert.strictEqual(block.nonce, f.nonce)
@@ -113,10 +116,16 @@ describe('Block', function () {
       it('returns ' + f.merkleRoot + ' for ' + f.id, function () {
         assert.strictEqual(Block.calculateMerkleRoot(block.transactions).toString('hex'), f.merkleRoot)
       })
+
+      if (f.witnessCommit) {
+        it('returns witness commit ' + f.witnessCommit + ' for ' + f.id, function () {
+          assert.strictEqual(Block.calculateMerkleRoot(block.transactions, true).toString('hex'), f.witnessCommit)
+        })
+      }
     })
   })
 
-  describe('checkMerkleRoot', function () {
+  describe('checkTxRoots', function () {
     fixtures.valid.forEach(function (f) {
       if (f.hex.length === 160) return
 
@@ -127,7 +136,7 @@ describe('Block', function () {
       })
 
       it('returns ' + f.valid + ' for ' + f.id, function () {
-        assert.strictEqual(block.checkMerkleRoot(), true)
+        assert.strictEqual(block.checkTxRoots(), true)
       })
     })
   })