+++ test/transaction.js
@@ -2,7 +2,7 @@ const { describe, it, beforeEach } = require('mocha')
 const assert = require('assert')
 const bscript = require('../src/script')
 const fixtures = require('./fixtures/transaction')
-const Transaction = require('../src/transaction')
+const Transaction = require('..').Transaction
 
 describe('Transaction', function () {
   function fromRaw (raw, noWitness) {