+++ ts_src/templates/pubkeyhash/output.ts
@@ -1,0 +1,20 @@
+// OP_DUP OP_HASH160 {pubKeyHash} OP_EQUALVERIFY OP_CHECKSIG
+
+import * as bscript from '../../script';
+import { OPS } from '../../script';
+
+export function check(script: Buffer | Array<number | Buffer>): boolean {
+  const buffer = bscript.compile(script);
+
+  return (
+    buffer.length === 25 &&
+    buffer[0] === OPS.OP_DUP &&
+    buffer[1] === OPS.OP_HASH160 &&
+    buffer[2] === 0x14 &&
+    buffer[23] === OPS.OP_EQUALVERIFY &&
+    buffer[24] === OPS.OP_CHECKSIG
+  );
+}
+check.toJSON = (): string => {
+  return 'pubKeyHash output';
+};