+++ types/bufferutils.d.ts
@@ -1,0 +1,4 @@
+/// <reference types="node" />
+export declare function readUInt64LE(buffer: Buffer, offset: number): number;
+export declare function writeUInt64LE(buffer: Buffer, value: number, offset: number): number;
+export declare function reverseBuffer(buffer: Buffer): Buffer;