+++ types/payments/lazy.d.ts
@@ -1,0 +1,2 @@
+export declare function prop(object: {}, name: string, f: () => any): void;
+export declare function value<T>(f: () => T): () => T;